Move touch functions to their own files
