+++ nano_gui.h
@@ -1,12 +1,6 @@
 #ifndef _NANO_GUI_H_
 #define _NANO_GUI_H_
 
-/* UI functions */
-struct Point {
-  int x, y;
-};
-extern struct Point ts_point;
-
 enum TextJustification_e : uint8_t
 {
   Left,
@@ -25,11 +19,6 @@ void displayChar(int16_t x, int16_t y, unsigned char c, uint16_t color, uint16_t
 void displayText(char *text, int x1, int y1, int w, int h, int color, int background, int border, TextJustification_e justification = TextJustification_e::Center);
 
 void formatFreq(uint32_t freq, char* buff, uint16_t buff_size, uint8_t fixed_width = 0);
-
-/* touch functions */
-bool readTouch(Point *const p);
-
-void scaleTouch(Point *const p);
 
 #define TEXT_LINE_HEIGHT 18
 #define TEXT_LINE_INDENT 5
